// src/solidity/type.ts
export interface ComplitionSources {
    [contractName: string]: CompilationSource;
}

export interface CompiledContractObj {
    /** If the language used has no contract names, this field should equal to an empty string. */
    [fileName: string]: {
        [contract: string]: CompiledContract;
    };
}

export interface CompilationResult {
    /** not present if no errors/warnings were encountered */
    errors?: CompilationError[];
    /** This contains the file-level outputs. In can be limited/filtered by the outputSelection settings */
    sources?: ComplitionSources;
    /** This contains the contract-level outputs. It can be limited/filtered by the outputSelection settings */
    contracts?: CompiledContractObj;
}
// ...
interface TypeDescription {
    typeIdentifier: string;
    typeString: string;
}

export enum AstNodeType {
    SourceUnit = 'SourceUnit',
    PragmaDirective = 'PragmaDirective',
    ImportDirective = 'ImportDirective',
    ContractDefinition = 'ContractDefinition',
    InheritanceSpecifier = 'InheritanceSpecifier',
    UsingForDirective = 'UsingForDirective',
    StructDefinition = 'StructDefinition',
    EnumDefinition = 'EnumDefinition',
    EnumValue = 'EnumValue',
    ParameterList = 'ParameterList',
    OverrideSpecifier = 'OverrideSpecifier',
    FunctionDefinition = 'FunctionDefinition',
    VariableDeclaration = 'VariableDeclaration',
    ModifierDefinition = 'ModifierDefinition',
    ModifierInvocation = 'ModifierInvocation',
    EventDefinition = 'EventDefinition',
    ElementaryTypeName = 'ElementaryTypeName',
    UserDefinedTypeName = 'UserDefinedTypeName',
    FunctionTypeName = 'FunctionTypeName',
    Mapping = 'Mapping',
    ArrayTypeName = 'ArrayTypeName',
    InlineAssembly = 'InlineAssembly',
    Block = 'Block',
    PlaceholderStatement = 'PlaceholderStatement',
    IfStatement = 'IfStatement',
    TryCatchClause = 'TryCatchClause',
    TryStatement = 'TryStatement',
    WhileStatement = 'WhileStatement',
    DoWhileStatement = 'DoWhileStatement',
    ForStatement = 'ForStatement',
    Continue = 'Continue',
    Break = 'Break',
    Return = 'Return',
    Throw = 'Throw',
    EmitStatement = 'EmitStatement',
    VariableDeclarationStatement = 'VariableDeclarationStatement',
    ExpressionStatement = 'ExpressionStatement',
    Conditional = 'Conditional',
    Assignment = 'Assignment',
    TupleExpression = 'TupleExpression',
    UnaryOperation = 'UnaryOperation',
    BinaryOperation = 'BinaryOperation',
    FunctionCall = 'FunctionCall',
    FunctionCallOptions = 'FunctionCallOptions',
    NewExpression = 'NewExpression',
    MemberAccess = 'MemberAccess',
    IndexAccess = 'IndexAccess',
    IndexRangeAccess = 'IndexRangeAccess',
    ElementaryTypeNameExpression = 'ElementaryTypeNameExpression',
    Literal = 'Literal',
    Identifier = 'Identifier',
    StructuredDocumentation = 'StructuredDocumentation',
    YulTypedName = 'YulTypedName',
    YulIdentifier = 'YulIdentifier',
    YulLiteral = 'YulLiteral',
    YulVariableDeclaration = 'YulVariableDeclaration',
    YulBlock = 'YulBlock'
}
// ...
export interface OverrideSpecifierAstNode {
    id: number;
    nodeType: AstNodeType.OverrideSpecifier;
    src: string;
    overrides: Array<UserDefinedTypeNameAstNode>;
}
// ...
export interface VariableDeclarationAstNode {
    id: number;
    nodeType: AstNodeType.VariableDeclaration;
    src: string;
    name: string;
    typeName: ElementaryTypeNameAstNode | UserDefinedTypeNameAstNode;
    constant: boolean;
    stateVariable: boolean;
    storageLocation: 'storage' | 'memory' | 'calldata' | 'default';
    overrides: OverrideSpecifierAstNode | null;
    visibility: string;
    value: string | null;
    scope: number;
    typeDescriptions: TypeDescription;
    functionSelector?: string;
    indexed?: boolean;
    baseFunctions?: Record<string, unknown>;
    mutability: string;
}
// ...
export interface ElementaryTypeNameAstNode {
    id: number;
    nodeType: AstNodeType.ElementaryTypeName;
    src: string;
    name: string;
    typeDescriptions: TypeDescription;
    stateMutability?: 'pure' | 'view' | 'nonpayable' | 'payable';
}

export interface UserDefinedTypeNameAstNode {
    id: number;
    nodeType: AstNodeType.UserDefinedTypeName;
    src: string;
    name: string;
    referencedDeclaration: number;
    contractScope: number | null;
    typeDescriptions: TypeDescription;
}
// ...
export interface CompilationError {
    /** Location within the source file */
    sourceLocation?: {
        file: string;
        start: number;
        end: number;
    };
    /** Error type */
    type?: CompilationErrorType;
    /** Component where the error originated, such as "general", "ewasm", etc. */
    component?: 'general' | 'ewasm' | string;
    severity?: 'error' | 'warning';
    message?: string;
    mode?: 'panic';
    /** the message formatted with source location */
    formattedMessage?: string;
}

type CompilationErrorType =
    | 'JSONError'
    | 'IOError'
    | 'ParserError'
    | 'DocstringParsingError'
    | 'SyntaxError'
    | 'DeclarationError'
    | 'TypeError'
    | 'UnimplementedFeatureError'
    | 'InternalCompilerError'
    | 'Exception'
    | 'CompilerError'
    | 'FatalError'
    | 'Warning';

/// /////////
// SOURCE //
/// /////////
export interface CompilationSource {
    /** Identifier of the source (used in source maps) */
    id: number;
    /** The AST object */
    ast: AstNode;
}

/// //////
// AST //
/// //////
export interface AstNode {
    absolutePath?: string;
    exportedSymbols?: Record<string, unknown>;
    id: number;
    nodeType: string;
    nodes?: Array<AstNode>;
    // format = s:l:f
    // 's': the byte-offset to the start of the range in the source file
    // 'l': the length of the source range in bytes
    // 'f': the source index mentioned above [CompilationSource.id]
    src: string;
    literals?: Array<string>;
    file?: string;
    scope?: number;
    sourceUnit?: number;
    symbolAliases?: Array<string>;
    [x: string]: any;
}

export interface AstNodeAtt {
    operator?: string;
    string?: null;
    type?: string;
    value?: string;
    constant?: boolean;
    name?: string;
    public?: boolean;
    exportedSymbols?: Record<string, unknown>;
    argumentTypes?: null;
    absolutePath?: string;
    [x: string]: any;
}

/// ///////////
// CONTRACT //
/// ///////////
export interface CompiledContract {
    /** The Ethereum Contract ABI. If empty, it is represented as an empty array. */
    abi: AbiItem[];
    // See the Metadata Output documentation (serialised JSON string)
    metadata: string;
    /** User documentation (natural specification) */
    userdoc: UserDocumentation;
    /** Developer documentation (natural specification) */
    devdoc: DeveloperDocumentation;
    /** Intermediate representation (string) */
    ir: string;
    /** EVM-related outputs */
    evm: {
        assembly: string;
        legacyAssembly: Record<string, unknown>;
        /** Bytecode and related details. */
        bytecode: BytecodeObject;
        deployedBytecode: BytecodeObject;
        /** The list of function hashes */
        methodIdentifiers: {
            [functionIdentifier: string]: string;
        };
        // Function gas estimates
        gasEstimates: {
            creation: {
                codeDepositCost: string;
                executionCost: 'infinite' | string;
                totalCost: 'infinite' | string;
            };
            external: {
                [functionIdentifier: string]: string;
            };
            internal: {
                [functionIdentifier: string]: 'infinite' | string;
            };
        };
    };
    /** eWASM related outputs */
    ewasm: {
        /** S-expressions format */
        wast: string;
        /** Binary format (hex string) */
        wasm: string;
    };
}